Introducing Artists Catalog View Model
