+++ TheMET/TheMET/ViewControllers/ArtistsCatalogViewController.swift
@@ -12,63 +12,64 @@ import MetAPI
 
 class ArtistsCatalogViewController: UIViewController {
     
-    private let artistsService = ArtistsService.standart
-    
-    private var artistsServiceSubscriber: AnyCancellable?
-    
-    private let metAPI = MetAPI()
+    private var viewModel: ArtistsCatalogViewModel?
     
     private let loadingArtistsCatalogView = LoadingPlaceholderView.construstView(configuration: .catalogLoading)
     private let failedArtitstCatalogView = FailedPlaceholderView.constructView(configuration: .catalogFailed)
     private let loadedArtistsCatalogView = FeaturedArtistsContentView.constructView()
     
-    private var loadedArtists: [FeaturedArtistsCellData] = []
+    private var loadedArtistsSubscriber: AnyCancellable?
+    private var contentStatusSubscriber: AnyCancellable?
     
-    private var loadingArtistNames: [String] = []
-    
-    var contentStatus:LoadingStatus<[FeaturedArtistsCellData]> = .loading {
-        didSet {
-            self.updateContent()
-        }
-    }
     
     override func viewDidLoad() {
         super.viewDidLoad()
         self.add(artistsCatalogSubview: self.failedArtitstCatalogView)
         self.add(artistsCatalogSubview: self.loadedArtistsCatalogView)
         self.add(artistsCatalogSubview: self.loadingArtistsCatalogView)
-        self.failedArtitstCatalogView.onButtonTap = { [weak self] in
-            self?.reloadButtonDidTap()
-        }
-        self.loadedArtistsCatalogView.onFeaturedArtistsCellTap = { [weak self] artistName in
-            self?.artistsCellDidTap(artistName)
-        }
-        self.loadedArtistsCatalogView.onFeaturedArtistsCellWillDisplay = { [weak self] artistName in
-            self?.loadParticularArtistCellDataIfNeeded(artistName: artistName)
-        }
-        self.artistsServiceSubscriber = self.artistsService.$artistsList
-            .sink(receiveValue: { [weak self] newArtistsList in
-                self?.reloadCatalog(artistsList: newArtistsList)
-            })
+        self.setupViewModel()
     }
     
-    private func updateContent() {
-        switch contentStatus {
-        case .failed:
-            self.loadedArtistsCatalogView.content = []
-            self.loadingArtistsCatalogView.isHidden = true
-            self.failedArtitstCatalogView.isHidden = false
-            self.loadedArtistsCatalogView.isHidden = true
-        case .loaded(let catalogCellDataList):
-            self.loadedArtistsCatalogView.content = catalogCellDataList
-            self.loadingArtistsCatalogView.isHidden = true
-            self.failedArtitstCatalogView.isHidden = true
-            self.loadedArtistsCatalogView.isHidden = false
-        case .loading:
-            self.loadedArtistsCatalogView.content = []
-            self.loadingArtistsCatalogView.isHidden = false
-            self.failedArtitstCatalogView.isHidden = true
-            self.loadedArtistsCatalogView.isHidden = true
+    private func setupViewModel() {
+        let viewModel = ArtistsCatalogViewModel { [weak self] in
+            return self
+        }
+        self.viewModel = viewModel
+        self.loadedArtistsSubscriber = viewModel.$loadedArtists
+            .sink(receiveValue: { [weak self] newLoadedArtists in
+                guard let self = self else {return}
+                self.loadedArtistsCatalogView.content = newLoadedArtists
+            })
+        self.contentStatusSubscriber = viewModel.$contentStatus
+            .sink(receiveValue: { [weak self] newContentStatus in
+                guard let self = self else {return}
+                switch newContentStatus {
+                case .failed:
+                    self.loadedArtistsCatalogView.content = []
+                    self.loadingArtistsCatalogView.isHidden = true
+                    self.failedArtitstCatalogView.isHidden = false
+                    self.loadedArtistsCatalogView.isHidden = true
+                case .loaded(let catalogCellDataList):
+                    self.loadedArtistsCatalogView.content = catalogCellDataList
+                    self.loadingArtistsCatalogView.isHidden = true
+                    self.failedArtitstCatalogView.isHidden = true
+                    self.loadedArtistsCatalogView.isHidden = false
+                case .loading:
+                    self.loadedArtistsCatalogView.content = []
+                    self.loadingArtistsCatalogView.isHidden = false
+                    self.failedArtitstCatalogView.isHidden = true
+                    self.loadedArtistsCatalogView.isHidden = true
+                }
+            })
+        self.failedArtitstCatalogView.onButtonTap = { [weak self] in
+            self?.viewModel?.reloadButtonDidTap()
+        }
+        self.loadedArtistsCatalogView.onFeaturedArtistsCellTap = { [weak self] artistName in
+            self?.viewModel?.artistsCellDidTap(artistName
+            )
+        }
+        self.loadedArtistsCatalogView.onFeaturedArtistsCellWillDisplay = { [weak self] artistName in
+            self?.viewModel?.loadParticularArtistCellDataIfNeeded(artistName: artistName)
         }
     }
     
@@ -83,110 +84,4 @@ class ArtistsCatalogViewController: UIViewController {
         ])
     }
     
-    private func reloadButtonDidTap() {
-        self.reloadCatalog(artistsList: self.artistsService.artistsList)
-    }
-    
-    private func reloadCatalog(artistsList: [String]) {
-        self.contentStatus = .loading
-        var artistsCellDataList: [FeaturedArtistsCellData] = []
-        for artist in artistsList {
-            let artistCellData: FeaturedArtistsCellData = FeaturedArtistsCellData(artistName: artist, artistData: .placeholder)
-            artistsCellDataList.append(artistCellData)
-        }
-        self.loadedArtists = artistsCellDataList
-        self.contentStatus = .loaded(artistsCellDataList)
-    }
-    
-    private func artistsCellDidTap(_ artistName: String) {
-        let artistArtsViewController = ArtistArtsViewController()
-        artistArtsViewController.artistName = artistName
-        self.navigationController?.pushViewController(artistArtsViewController, animated: true)
-    }
-    
-    private func loadParticularArtistCellDataIfNeeded(artistName: String) {
-        guard case .loaded(let artistCellDataList) = self.contentStatus else {
-            return
-        }
-        guard let artistCellData = artistCellDataList.first(where: { cellData in
-            cellData.artistName == artistName
-        }),
-              case .placeholder = artistCellData.artistData,
-              !self.loadingArtistNames.contains(artistName),
-              let artist = self.loadedArtists.first(where: { artist in
-                  artist.artistName == artistName
-              })
-        else { return }
-        self.loadingArtistNames.append(artist.artistName)
-        self.loadArtistCellData(artistName: artistName, completion: { [weak self] artistCellData in
-            guard let self = self else { return }
-            if case .loaded(var artistsCellDataList) = self.contentStatus,
-               let artistCellDataIndex = artistsCellDataList.firstIndex(where: { $0.artistName == artistCellData.artistName}) {
-                   artistsCellDataList[artistCellDataIndex] = artistCellData
-                self.contentStatus = .loaded(artistsCellDataList)
-            }
-            self.loadingArtistNames.removeAll { name in
-                return name == artistName
-            }
-        })
-    }
-    
-    private func loadArtistCellData(artistName: String, attempt: Int = 0, completion: @escaping (FeaturedArtistsCellData) -> Void) {
-        let parameters:[SearchParameter] = [
-            .artistOrCulture(true),
-            .q(artistName)
-        ]
-        self.metAPI.search(parameters: parameters) { [weak self] searchResults in
-            guard let self = self else {
-                return
-            }
-            switch searchResults {
-            case .failure(let error):
-                guard case .nonDecodableData = error,
-                      attempt > 5 else {
-                    self.loadArtistCellData(artistName: artistName, attempt: attempt + 1, completion: completion)
-                    return
-                }
-                let artistCellData = FeaturedArtistsCellData(artistName: artistName, artistData: .data(imageURL: nil, title: artistName, subTitle: nil))
-                completion(artistCellData)
-            case .success(let objects):
-                self.loadArtistImageURL(objectsIDs: objects.objectIDs) { url in
-                    let artistCellData = FeaturedArtistsCellData(artistName: artistName, artistData: .data(imageURL: url, title: artistName, subTitle: nil))
-                    completion(artistCellData)
-                }
-            }
-        }
-    }
-
-    private func loadArtistImageURL(objectsIDs: [ArtID], completion: @escaping (URL?) -> Void) {
-        guard !objectsIDs.isEmpty else {
-            completion(nil)
-            return
-        }
-        self.loadObjectImageURL(objectId: objectsIDs[0]) { [weak self] url in
-            if let imageURL = url {
-                completion(imageURL)
-            } else {
-                let newObjectsIds: [ArtID] = Array(objectsIDs.dropFirst())
-                self?.loadArtistImageURL(objectsIDs: newObjectsIds, completion: completion)
-            }
-        }
-    }
-    
-    private func loadObjectImageURL(objectId: ArtID, completion: @escaping (URL?) -> Void){
-        self.metAPI.object(id: objectId) { objectResult in
-            switch objectResult {
-            case .failure:
-                completion(nil)
-            case .success(let object):
-                guard let imageURLstring = object.primaryImage else {
-                    completion(nil)
-                    return
-                }
-                completion(URL(string: imageURLstring))
-            }
-        }
-    }
-
-
 }
